requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2
django-dotenv==1.4.2
django-filter==24.2
djangorestframework==3.15.2
idna==3.7
importlib_metadata==7.2.0
Markdown==3.6
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
